ASKAPSDP-1638 - Getting the component catalogue flux units correct.
As noted, the components catalogue had fluxes that were not in mJy or 
mJy/beam, but rather in the image flux units (ie. Jy/beam etc). This 
commit fixes that, by fixing the column units to mJy-based, and then 
converting from the image flux units to mJy or mJy/beam for the flux 
entries.

Note that the islands catalogue has *not* been converted, as it uses 
the Duchamp code to write the entries. 


git-svn-id: https://svn.atnf.csiro.au/askapsdp/trunk@6735 1db04d4c-0ff0-4329-ae0b-98c546476c6a