ASKAPSDP-1638 - Improving the code interface for writing out island and component catalogues
I have made a number of improvements to the way we handle the writing
of the island and component tables. There are now new classes for
island and component objects, and for catalogues thereof - all located
in the new catalogues directory. The new island/component classes
instantiate directly from a RadioSource object that contains all the
data, so we initialise with the correct information once.
We have the specification defined in the one place, along with the
catalogue version information, and this works a bit better with the
output classes, which have been simplified a bit as a result.
We also now have .txt versions of the catalogues produced. The old
fitResults files are also produced in the same way, using the
component catalogue, and the annotation files are now produced via the
catalogue interface as well.
There is still a bit of tidying up needed, particularly in the
RadioSource class, but it's OK for now.
And I've added much more documentation in the .h files!


git-svn-id: https://svn.atnf.csiro.au/askapsdp/trunk@6753 1db04d4c-0ff0-4329-ae0b-98c546476c6a