ASKAPSDP-1654 - Fixing the positional offsets for the component parset output.
This change replaces the old way of finding the RA/DEC offsets (which 
used great-circle displacements) with the better way of getting the 
projected offsets. This replicates the code found in 
gridding/AWProjectVisGridder.cc in Synthesis. Tests with the CDFS BETA 
data indicate an improvement in the output of calibration.


git-svn-id: https://svn.atnf.csiro.au/askapsdp/trunk@6760 1db04d4c-0ff0-4329-ae0b-98c546476c6a